Admins can bypass webclient block
There is always the DD dropdown to flat out disable the webclient

+++ code/modules/client/client_procs.dm
@@ -167,7 +167,7 @@ var/next_external_rsc = 0
 		src << "Required version to remove this message: [config.client_warn_version] or later"
 		src << "Visit http://www.byond.com/download/ to get the latest version of byond."
 
-	if (connection == "web")
+	if (connection == "web" && !holder)
 		if (!config.allowwebclient)
 			src << "Web client is disabled"
 			qdel(src)
